MNK-3448 : check for null return (happens when directory does not exist)
git-svn-id: http://svn.terracotta.org/svn/ehcache@5296 b9324663-ca0f-0410-8574-be9b3887307d

+++ trunk/ehcache/ehcache-core/src/test/java/net/sf/ehcache/constructs/blocking/BlockingCacheTest.java
@@ -56,10 +56,16 @@ import static org.junit.Assert.fail;
  */
 public final class BlockingCacheTest {
 
+    private static final String DISK_STORE_PATH = "target/BlockingCacheTest";
+    
     @BeforeClass
     public static void cleanupDisk() {
-        for (File f : new File("target/BlockingCacheTest").listFiles()) {
-            f.delete();
+        File diskStorePath = new File(DISK_STORE_PATH);
+        File[] files = diskStorePath.listFiles();
+        if (files != null) {
+            for (File f : files) {
+                f.delete();
+            }
         }
     }
     
@@ -86,7 +92,7 @@ public final class BlockingCacheTest {
     private CacheManager createCacheManager(CacheConfiguration config) {
         CacheManager manager = new CacheManager(new Configuration()
                 .name("BlockingCacheTest")
-                .diskStore(new DiskStoreConfiguration().path("target/BlockingCacheTest")));
+                .diskStore(new DiskStoreConfiguration().path(DISK_STORE_PATH)));
         manager.addCache(new Cache(config));
         return manager;
     }
